LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-align-baseline-002-expected.txt

line1
line2
line1
line2
line1
line2
line1
line2
line1
line2
line1
line2

PASS #target > div 1
PASS #target > div 2
FAIL #target > div 3 assert_equals:
<div data-offset-x="126">line1<br>line2</div>
offsetLeft expected 126 but got 125
PASS #target > div 4
PASS #target > div 5
PASS #target > div 6
